pyproject.toml
[project]
name = "mypdfwrapper"
version = "0.1.0"
description = "PDF extraction wrapper built on PyMuPDF"
requires-python = ">=3.10"
dependencies = [
    "pymupdf>=1.24.0",
    "python-dotenv",
    "pytesseract",
    "pillow"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.hatch.build.targets.wheel]
packages = ["src/mypdfwrapper"]
